requirements.txt
alembic==1.13.1
blinker==1.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.3.0
Werkzeug==3.0.2
WTForms==3.1.2
